feat(workflow): node catalog skeleton and the variable-inference split
Phase 1 steps 0+1 of the node-catalog migration. No node type migrates yet;
no node panel, body, or trace renderer changes by a byte.

Step 0 - the split that unblocks everything: utils/variable-utils.ts mixed
pure variable-id/shape inference with store-reading display helpers, and
because variable-conversion.ts imported two of the pure ones, ~24 node
schema files transitively pulled zustand plus a React barrel. The pure layer
(buildVariableId, getLabelFromVariableId, containsVariableReference,
isNodeVariable, isVariableMode, inferPluckOutputType, the array-segment
helpers, and friends) now lives in lib
(workflow-engine/catalog/variable-inference.ts); UnifiedVariable,
AllowedVarType and ValidationRules move with it
(workflow-engine/types/unified-variable.ts). variable-conversion.ts now has
deliberately zero ~/ imports; the web variable-utils keeps only the six
useResourceStore display helpers and re-exports the moved names so none of
its ~26 importers churn.

Step 1 - the catalog skeleton: catalog/types.ts (NodeManifest - the data
half of a node definition, with an enforced configSchema, agentSchema
projection slot, connection rules incl. the branches() contract, and agent
docs; NodeCategory and the builder's validator result relocate here, the
latter as NodeValidationResult to avoid the engine's ValidationResult),
catalog/registry.ts (module-level map, no class), and
catalog/not-yet-migrated.ts listing all 38 NodeType values. apps/web gains
defineFromManifest, which builds today's NodeDefinition from a manifest plus
the React parts, and a coverage test asserting every NodeType is in exactly
one of {manifests, NOT_YET_MIGRATED} plus configSchema.safeParse(defaults)
for every registered manifest - the discipline the legacy schema:any never
had. Output resolution stays deferred to the next phase (commented slot).
